ocean/requirements.txt
torch>=2.0.0
numpy>=1.21.0
scikit-learn>=1.0.0
pandas>=1.3.0
networkx>=2.6.0
matplotlib>=3.4.0
tqdm>=4.62.0
pydantic>=2.0.0
pyyaml>=6.0.0
requests>=2.25.0
mypy>=1.0.0
pytest>=7.0.0
black>=22.0.0
isort>=5.0.0
flake8>=4.0.0
python-dotenv>=0.19.0
loguru>=0.6.0
structlog>=22.0.0
fastapi>=0.68.0
uvicorn>=0.15.0
gunicorn>=20.1.0
python-multipart>=0.0.5